Move lib into src directory. With this it'll do the babel transformation properly.
